Plumbing to create save states
Create serialise.c. window.c is unfortunately left more stateful than it
was before, but it doesn't seem sensible to generalise the input wiring
before we have more methods to apply it to. vm.c is also not beautiful
as a result of the need to save at a specific point in the loop.